% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_clonotypes.R
\name{cluster_clonotypes}
\alias{cluster_clonotypes}
\title{Cluster Clonotypes By Distribution}
\usage{
cluster_clonotypes(
  seurat_object,
  clonotype_column,
  cell_type_column,
  sample_column,
  clone_min = 5,
  plot = TRUE
)
}
\arguments{
\item{seurat_object}{seurat object to operate on}

\item{clonotype_column}{Character describing name of column to use for clone calling}

\item{sample_column}{Character describing name of column that specifies samples to distinguish ontogically unrelated but potentially identical clonotypes}

\item{clone_min}{Integer specifying minimum number fo clones for a clonotype to be included in analysis}

\item{plot}{Boolean specifying whether or not to plot a clustered heatmap}

\item{celltype_column}{Character describing name of column to cluster clonotypes across}
}
\value{
a dataframe with clonotypes as rows and their propotional distribution across celltypes (or the variables contained within 'celltype_column') as columns
}
\description{
This function plots a UMAP where cells can be colored according to a variable.
Multiple clonotypes will also be highlighted.
}
