% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_clonotype.R
\name{highlight_clonotype}
\alias{highlight_clonotype}
\title{Highlighting a Single Clonotype}
\usage{
highlight_clonotype(
  seurat_object,
  clonotype,
  clonotype_column,
  group.by,
  highlight_color = "black",
  other_colors,
  highlight_size = 1,
  other_size = 0.3,
  other_alpha = 0.1
)
}
\arguments{
\item{seurat_object}{seurat object to operate on}

\item{clonotype}{Character specifying a clonotype within the 'clonotype_column'}

\item{clonotype_column}{Character describing name of column to use for clone calling}

\item{group.by}{Character describing a column to use for the coloring of cells}

\item{highlight_color}{Color of the highlighted clonotype}

\item{other_colors}{Vector of colors to use for coloring groups in 'group.by'}

\item{highlight_size}{Number representing the point size of the highlighted clonotype}

\item{other_size}{Number representing the point size of the non-highlighted cells}

\item{other_alpha}{Number (0-1) corresponding to the transparency of non-highlighted cells}
}
\value{
a ggplot object that is plotted and can be stored/modified
}
\description{
This function plots a UMAP where cells can be colored according to a variable.
A single clonotype will also be highlighted.
}
