% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_multiple_clonotypes.R
\name{highlight_multiple_clonotypes}
\alias{highlight_multiple_clonotypes}
\title{Highlighting Multiple Clonotypes}
\usage{
highlight_multiple_clonotypes(
  seurat_object,
  clonotypes,
  clonotype_column,
  highlight_colors,
  other_color = "lightgray",
  highlight_size = 1,
  other_size = 0.3,
  other_alpha = 0.1
)
}
\arguments{
\item{seurat_object}{seurat object to operate on}

\item{clonotypes}{Character vector specifying clonotypes within the 'clonotype_column'}

\item{clonotype_column}{Character describing name of column to use for clone calling}

\item{highlight_colors}{Vector of colors to use for the highlighted clonotypes}

\item{other_color}{Color to use for the non-highlighted cells}

\item{highlight_size}{Number representing the point size of the highlighted clonotypes}

\item{other_size}{Number representing the point size of the non-highlighted cells}

\item{other_alpha}{Number (0-1) corresponding to the transparency of non-highlighted cells}
}
\value{
a ggplot object that is plotted and can be stored/modified
}
\description{
This function plots a UMAP where cells can be colored according to a variable.
Multiple clonotypes will also be highlighted.
}
