% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{Plot Clonal Frequencies}
\usage{
plot_freq(
  seurat_object,
  clonotypes,
  samples,
  cell_type_order,
  cell_type_colors,
  plot.type = "alluvial"
)
}
\arguments{
\item{seurat_object}{seurat object to operate on}

\item{sample.1}{Character describing a sample in 'sample_column' for comparison}

\item{sample.2}{Character describing a sample in 'sample_column' for comparison}

\item{minimum_count}{integer value specifying the minimum number of clones for a clonotype to be included in analysis}

\item{decreasing}{Boolean value specifying whether or not to sort by decreasing (or increasing) expansion}

\item{identity}{A value in 'clonotype_identity_mode' to specify clonotypes of interest (must first run the 'calculate_freq' function)}
}
\value{
a ggplot object that is plotted and can be stored/modified.
}
\description{
This function takes a seurat object with calculated frequencies and a list of clonotypes
and samples. You can specify either an alluvial or a circular barplot form and the function
plot clonal frequencies and phenotypes across multiple samples.
}
