% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_aesthetics.R
\name{plotting_aesthetics}
\alias{plotting_aesthetics}
\title{Plotting Aesthetics}
\usage{
plotting_aesthetics(seurat_object, cell_type_order, cell_type_colors)
}
\arguments{
\item{seurat_object}{seurat object to operate on}

\item{cell_type_order}{vector that includes cell types in desired order}

\item{cell_type_colors}{vector that includes colors in corresponding order with 'cell_type_order'}
}
\value{
a seurat object with new stored plotting aesthetics.
}
\description{
This function takes a stores plotting aesthetics in a seurat object for simple, reproducible plotting.
}
